+++ internal/httpapi/assets/_app/immutable/chunks/C9HyNfF7.js
@@ -1,1 +1,0 @@
-import"./DwJOBqur.js";import{s,r as p}from"./m6oBLQyM.js";import{I as a}from"./q5H6Yfri.js";var n=new Set(["$$slots","$$events","$$legacy"]);function i(r,e){let o=p(e,n);const t=[["path",{d:"m12 19-7-7 7-7"}],["path",{d:"M19 12H5"}]];a(r,s({name:"arrow-left"},()=>o,{get iconNode(){return t}}))}export{i as A};