+++ internal/httpapi/assets/_app/immutable/chunks/CLGSeIC2.js
@@ -1,1 +1,0 @@
-import"./DwJOBqur.js";import{s,r as p}from"./m6oBLQyM.js";import{I as n}from"./q5H6Yfri.js";var a=new Set(["$$slots","$$events","$$legacy"]);function i(e,o){let r=p(o,a);const t=[["path",{d:"M18 6 6 18"}],["path",{d:"m6 6 12 12"}]];n(e,s({name:"x"},()=>r,{get iconNode(){return t}}))}export{i as X};