+++ internal/httpapi/assets/_app/immutable/chunks/cVoeTi1q.js
@@ -1,1 +1,0 @@
-import"./DwJOBqur.js";import{s,r as n}from"./m6oBLQyM.js";import{I as p}from"./q5H6Yfri.js";var a=new Set(["$$slots","$$events","$$legacy"]);function $(r,e){let o=n(e,a);const t=[["path",{d:"m9 18 6-6-6-6"}]];p(r,s({name:"chevron-right"},()=>o,{get iconNode(){return t}}))}export{$ as C};